DC requested changes / Merged manage-account-info for all users into a shared page
