--- src/modules/feature-modules/accessor/pages/account/manage-account-info.component.html
@@ -1,13 +1,0 @@
-<theme-content-wrapper [status]="pageStatus">
-  <div class="nhsuk-grid-row">
-    <div class="nhsuk-grid-column-full">
-      <h2 class="nhsuk-heading-l">Change your Password</h2>
-      <p>You will need to enter your current password to change to a new one.</p>
-      <a [href]="changePassword"> Change your Password </a>
-      <h6>Last password update: {{ user.passwordResetAt | date: "dd/MM/yyyy" }}</h6>
-    </div>
-  </div>
-  <ng-container *ngIf="pageStatus === 'READY'">
-    <app-account-mfa-list [MFAInfo]="MFAInfo" />
-  </ng-container>
-</theme-content-wrapper>
