+++ src/modules/feature-modules/admin/admin.module.ts
@@ -9,8 +9,7 @@ import { AdminRoutingModule } from './admin-routing.module';
 import { ContextInnovationOutletComponent } from './base/context-innovation-outlet.component';
 import { SidebarInnovationMenuOutletComponent } from './base/sidebar-innovation-menu-outlet.component';
 // Pages.
-// // Account.
-import { PageAccountManageAccountInfoComponent } from './pages/account/manage-account-info.component';
+
 // // Users
 import { PageUsersRoleActivateComponent } from './pages/users/roles/role-activate.component';
 import { PageUsersRoleChangeComponent } from './pages/users/roles/role-change.component';
@@ -65,7 +64,6 @@ import { OrganisationDataResolver } from './resolvers/organisation-data.resolver
 import { OrganisationUnitDataResolver } from './resolvers/organisation-unit-data.resolver';
 import { ServiceUserDataResolver } from './resolvers/service-user-data.resolver';
 import { SidebarAccountMenuOutletComponent } from './base/sidebar-account-menu-outlet.component';
-import { PageAccountInfoComponent } from './pages/account/account-info.component';
 
 @NgModule({
   imports: [ThemeModule, SharedModule, AdminRoutingModule],
@@ -76,8 +74,7 @@ import { PageAccountInfoComponent } from './pages/account/account-info.component
     SidebarAccountMenuOutletComponent,
 
     // Pages.
-    PageAccountManageAccountInfoComponent,
-    PageAccountInfoComponent,
+
     // Users
     PageUserEmailComponent,
     PageUserFindComponent,
