--- src/modules/feature-modules/innovator/pages/account/account-info.component.html
@@ -1,22 +1,0 @@
-<theme-content-wrapper [status]="pageStatus">
-  <div class="nhsuk-grid-row nhsuk-u-margin-bottom-5">
-    <div class="nhsuk-grid-column-full">
-      <h2 class="nhsuk-heading-l">Change your Password</h2>
-      <p>You will need to enter your current password to change to a new one.</p>
-      <a [href]="changePassword"> Change your Password </a>
-      <h6>Last password update: {{ user.passwordResetAt | date: "dd/MM/yyyy" }}</h6>
-    </div>
-  </div>
-
-  <ng-container *ngIf="pageStatus === 'READY'">
-    <app-account-mfa-list [MFAInfo]="MFAInfo" />
-  </ng-container>
-
-  <div class="nhsuk-grid-row nhsuk-u-margin-bottom-5">
-    <div class="nhsuk-grid-column-full">
-      <h2 class="nhsuk-heading-l">Delete your account</h2>
-      <p>You will need to confirm your email address to delete your account.</p>
-      <a routerLink="delete"> Delete your account </a>
-    </div>
-  </div>
-</theme-content-wrapper>
