+++ src/modules/shared/pages/account/manage-account-info/manage-account-info.component.html
@@ -8,7 +8,15 @@
     </div>
   </div>
 
-  <ng-container *ngIf="pageStatus === 'READY'">
-    <app-account-mfa-list [MFAInfo]="MFAInfo" />
+  <app-account-mfa-list [MFAInfo]="MFAInfo" />
+
+  <ng-container *ngIf="isInnovator">
+    <div class="nhsuk-grid-row nhsuk-u-margin-bottom-5">
+      <div class="nhsuk-grid-column-full">
+        <h2 class="nhsuk-heading-l">Delete your account</h2>
+        <p>You will need to confirm your email address to delete your account.</p>
+        <a routerLink="delete"> Delete your account </a>
+      </div>
+    </div>
   </ng-container>
 </theme-content-wrapper>
