+++ src/modules/shared/pages/account/manage-account-info/manage-account-info.component.ts
@@ -1,20 +1,24 @@
 import { Component, OnInit } from '@angular/core';
+
 import { CoreComponent } from '@app/base';
+import { DateISOType } from '@app/base/types';
 import { AuthenticationService } from '@modules/stores';
 import { MFAInfoDTO } from '@modules/stores/authentication/authentication.service';
 
 @Component({
-  selector: 'app-assessment-pages-manage-account-info',
+  selector: 'shared-pages-account-manage-account-info',
   templateUrl: './manage-account-info.component.html'
 })
-export class PageAssessmentAccountManageAccountInfoComponent extends CoreComponent implements OnInit {
+export class PageSharedAccountManageAccountInfoComponent extends CoreComponent implements OnInit {
   changePassword = `${this.CONSTANTS.APP_URL}/change-password`;
 
   user: {
-    passwordResetAt: null | string;
+    passwordResetAt: null | DateISOType;
   };
 
   MFAInfo: MFAInfoDTO = { type: 'none' };
+
+  isInnovator: boolean = this.stores.authentication.isInnovatorType();
 
   constructor(private authenticationService: AuthenticationService) {
     super();
