+++ src/modules/stores/authentication/authentication.service.ts
@@ -39,7 +39,7 @@ type GetUserInfoDTO = {
   }[];
 };
 
-export type MFAInfoDTO = { type: 'none' } | { type: 'email' } | { type: 'phone'; phoneNumber: string };
+export type MFAInfoDTO = { type: 'none' } | { type: 'email' } | { type: 'phone'; phoneNumber: string | undefined };
 
 export type UpdateUserInfoDTO = {
   displayName: string;
@@ -150,9 +150,9 @@ export class AuthenticationService {
     );
   }
 
-  updateUserMFAInfo(body: MFAInfoDTO): Observable<{ id: string }> {
+  updateUserMFAInfo(body: MFAInfoDTO): Observable<void> {
     const url = new UrlModel(this.API_USERS_URL).addPath('v1/me/mfa');
-    return this.http.put<{ id: string }>(url.buildUrl(), body).pipe(
+    return this.http.put<void>(url.buildUrl(), body).pipe(
       take(1),
       map(response => response)
     );
